refactor: Condense `password` related files in one
